requirements.txt
python-telegram-bot==21.10
sqlalchemy[asyncio]==2.0.40
asyncpg>=0.30.0
python-dotenv==1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
